Build fix for the CPP example.
NotificationActivationCallback.h is not found because
the target platform is incorrectly put to 8.1 if you have
have the 8.1 SDK installed. It will not build as the header
is not included in that SDK. Enforcing to 10+ will make it compile.

Also renamed the include to follow
https://msdn.microsoft.com/en-us/library/windows/desktop/mt643711(v=vs.85).aspx
which uses a camelcase name : NotificationActivationCallback.h